Tool version cleanup and simplify gem installation of tmuxinator and bundler
Removing ruby, elixir and erlang as version we care about. Realistically
I want all tools to be installed globally and having different tool
versions (this project vs globally) was causing issues.
Sometimes the local ruby version in the project `.tool-versions` file
was out of sync with the global ruby version and this was causing
tmuxinator to be unavailable globally.