feat(tmux, nvim): remove nvim-tmux-navigation due to bug
There was an issue with the ctrl-h and ctrl-l key bindings were causing
issues in neovim. This keybinding is for navigating between neovim
splits (and additionally from tmux to neovim and vice versa). However
this started to cause neovim to slow to a crawl and become unusable with
multiple keys strikes required for each character to appear.