fix(nvim): mason-lspconfig 2.0 broke existing LSP configs
Neovim 0.11 simplified how LSPs could be configured, this is now
reflected in Mason-lspconfig which introduced a breaking 2.0 version
that removed quite a bit of existing functionality.

Mason-lspconfig is now used to install and configure only ESLint and
TypeScript LSPs. Custom LSP configurations are defined in `/lsp/*.lua`
and enabled in `lsp.lua` (a new file).
The risk with this is the the LSP configurations defined in `lsp/*.lua`
will not be kept upto date with the latest changes in the nvim-lspconfig
repository.
There is probably a happy path where mason-lspconfig is made redundant
by neovim LSP features.

See
https://github.com/mason-org/mason-lspconfig.nvim/blob/main/CHANGELOG.md#200-2025-05-06